moving test_early_outbreak_is_exponential to devtest_hiv.py
