+++ tests/devtests/devtest_hiv.py
@@ -5,6 +5,17 @@ import stisim as sti
 import hivsim
 import matplotlib.pyplot as plt
 
+from statistics import log
+from pathlib import Path
+import sys
+
+tests_directory = Path(__file__).resolve().parent.parent  # up to stisim/tests/ dir
+sys.path.append(str(tests_directory))
+
+from testlib import build_testing_sim
+
+verbose = False
+
 
 class TrackValues(ss.Analyzer):
     # Track outputs for viral load and CD4 counts
@@ -141,6 +152,7 @@ class PerformTest(ss.Intervention):
             self.sim.diseases.syphilis.set_prognoses(ss.uids(self.syphilis_infections[ti]))
 
 
+@sc.timer()
 def test_hivsim():
     """ Basic HIVsim test """
     sim = hivsim.Sim()
@@ -148,6 +160,7 @@ def test_hivsim():
     return sim
 
 
+@sc.timer()
 def test_hiv():
     """
     Legacy HIV test -- should deprecate
@@ -184,6 +197,7 @@ def test_hiv():
     return sim
 
 
+@sc.timer()
 def test_hiv_syph():
     """
     Legacy HIV-syphilis integration test
@@ -223,7 +237,56 @@ def test_hiv_syph():
     return sim
 
 
+@sc.timer()
+def test_early_outbreak_is_exponential():
+    sc.heading("Ensuring that early sexual-transmitted HIV outbreaks follow an exponential pattern.")
+
+    # short sim, no death/maternal/pregnancy to keep things simple
+
+    n_passing = 0
+    n_seeds = 10
+    all_sims = []
+    for rand_seed in range(n_seeds):
+        disease = sti.HIV(beta_m2f=0.02, beta_m2c=0, init_prev=0.01)
+        sim = build_testing_sim(n_agents=7000, duration=1, diseases=[disease],
+                                pregnancy=None, maternal_network=None, death=None)
+        sim.pars['rand_seed'] = rand_seed
+        sim.pars['verbose'] = False
+        all_sims.append(sim)
+
+        sim.run()
+
+        hiv = sim.diseases.hiv
+        n_infected = hiv.results.n_infected
+        log_n_infected = [log(n_infected[i]) for i in range(len(n_infected))]
+
+        # determining how much variance of log_n_infected is explained by a linear trend (linear trending log data => source is exponential)
+        corr_coef = np.corrcoef(range(len(log_n_infected)), log_n_infected)[0, 1]
+
+        if verbose:
+            print(f"log(n_infections) correlation with time: {corr_coef}")
+
+        threshold = 0.9  # correlation coefficient, threshold**2 is variance explained minimum
+        if (corr_coef > threshold):
+            n_passing += 1
+        else:
+            print(f"index/seed: {rand_seed} FAIL correlation coefficient test: {corr_coef} < {threshold}")
+
+    print(f"---\nExponential outbreak (passing) sims: {n_passing} fraction: {n_passing/n_seeds}\n---")
+
+    return all_sims
+
+
+
 if __name__ == '__main__':
+    do_plot = True
+    sc.options(interactive=do_plot)
+    timer = sc.timer()
+
     s1 = test_hivsim()
     s2 = test_hiv()
     s3 = test_hiv_syph()
+    sims = test_early_outbreak_is_exponential()
+
+    sc.heading("Total:")
+    timer.toc()
