+++ tests/test_hiv_natural_history_verification.py
@@ -12,7 +12,7 @@ import sys
 
 from itertools import chain
 from pathlib import Path
-from statistics import log, mean
+from statistics import mean
 
 import starsim as ss
 import stisim as sti
@@ -318,46 +318,6 @@ def test_no_hiv_with_no_outbreaks():
     return sim
 
 
-@sc.timer()
-def test_early_outbreak_is_exponential():
-    sc.heading("Ensuring that early sexual-transmitted HIV outbreaks follow an exponential pattern.")
-
-    # short sim, no death/maternal/pregnancy to keep things simple
-
-    n_passing = 0
-    n_seeds = 10
-    all_sims = []
-    for rand_seed in range(n_seeds):
-        disease = sti.HIV(beta_m2f=0.02, beta_m2c=0, init_prev=0.01)
-        sim = build_testing_sim(n_agents=7000, duration=1, diseases=[disease],
-                                pregnancy=None, maternal_network=None, death=None)
-        sim.pars['rand_seed'] = rand_seed
-        sim.pars['verbose'] = False
-        all_sims.append(sim)
-
-        sim.run()
-
-        hiv = sim.diseases.hiv
-        n_infected = hiv.results.n_infected
-        log_n_infected = [log(n_infected[i]) for i in range(len(n_infected))]
-
-        # determining how much variance of log_n_infected is explained by a linear trend (linear trending log data => source is exponential)
-        corr_coef = np.corrcoef(range(len(log_n_infected)), log_n_infected)[0, 1]
-
-        if verbose:
-            print(f"log(n_infections) correlation with time: {corr_coef}")
-
-        threshold = 0.9  # correlation coefficient, threshold**2 is variance explained minimum
-        if (corr_coef > threshold):
-            n_passing += 1
-        else:
-            print(f"index/seed: {rand_seed} FAIL correlation coefficient test: {corr_coef} < {threshold}")
-
-    print(f"---\nExponential outbreak (passing) sims: {n_passing} fraction: {n_passing/n_seeds}\n---")
-
-    return all_sims
-
-
 # Not currently implemented in hivsim, so leaving this partially-completed test commented out for future work
 # def test_perinatally_infected_progress_faster(self):
 #     sim = build_testing_sim(diseases=self.diseases, demographics=self.demographics,
@@ -389,7 +349,6 @@ if __name__ == '__main__':
     test_cd4_rises_on_ART()
     test_art_increases_longevity()
     test_no_hiv_with_no_outbreaks()
-    test_early_outbreak_is_exponential()
 
     sc.heading("Total:")
     timer.toc()
